+++ client/java/generator/src/main/java/io/openlineage/client/TypeResolver.java
@@ -35,6 +35,7 @@ import io.openlineage.client.SchemaParser.PrimitiveType;
 import io.openlineage.client.SchemaParser.RefType;
 import io.openlineage.client.SchemaParser.Type;
 import io.openlineage.client.SchemaParser.TypeVisitor;
+import java.util.stream.Collectors;
 
 /**
  * Resolves the types in a Schema. (ref, etc)
@@ -137,7 +138,15 @@ public class TypeResolver {
 
         @Override
         public ResolvedType visit(OneOfType oneOfType) {
-          throw new UnsupportedOperationException("oneOf is not supported yet " + oneOfType);
+          List<ResolvedType> resolvedTypes = oneOfType
+              .getTypes()
+              .stream()
+              .filter(type -> type instanceof RefType)
+              .map(refType -> visit(refType))
+              .collect(Collectors.toList());
+
+          // return first of
+          return resolvedTypes.get(0);
         }
 
         @Override